+++ Runtime/Configuration/ConfigurationUtils.cs
@@ -1,3 +1,5 @@
+using Unity.Services.Core.Internal.Serialization;
+
 namespace Unity.Services.Core.Configuration
 {
     static class ConfigurationUtils
@@ -5,6 +7,6 @@ namespace Unity.Services.Core.Configuration
         public const string ConfigFileName = "UnityServicesProjectConfiguration.json";
 
         public static IConfigurationLoader ConfigurationLoader { get; internal set; }
-            = new StreamingAssetsConfigurationLoader();
+            = new StreamingAssetsConfigurationLoader(new NewtonsoftSerializer());
     }
 }