+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.9.0] - 2023-05-09
+
+### Added
+
+- New common error codes `ProjectPolicyAccessDenied` and `PlayerPolicyAccessDenied`.
+
+### Changed
+
+- Newtonsoft (de)serialization calls to be thread safe.
+- Bumped Newtonsoft dependency to 3.2.1 to use the latest stable version.
+
 ## [1.8.2] - 2023-04-04
 
 ### Fixed